+++ lib/ui/users/expediteur/marchandises/annonces/plus.dart
@@ -67,7 +67,7 @@ class _NewAnnonceState extends State<NewAnnonce> {
     String nom = provider.nom;
     int tarif = calculatrice.montant;
     int quantite = provider.quantite;
-    int poids = provider.poids;
+    double poids = provider.poids;
     String adress_exp = provider.adress_exp;
     String adress_liv = provider.adress_liv;
     Villes ville_exp = provider.ville_exp;
